Move test_args_title to test_valid_args_schemas.py
